refactor(auth): create sendEmail function in utils
